requirements.txt
discord.py>=2.6.4
psutil>=7.2.1
python-dotenv>=1.2.1
sqlalchemy>=2.1.0b1